FR translation
Very quick french translation. 

While I tried to stick to the literal meaning of things sometimes it would be non-understandable by a native speaker, especially if new to the emulator scene. 

This should be tested with the french UI since the place of some links and words have been inverted and I don't know how it'll behave once in the UI.